Updating queries to expand record fields dynamically
The Airtable API does not return a field in the record object if the field has no value. The previous query excluded the field (column) altogether if the first record in the list did not contain the field name.

This query will dynamically build the columns based on all fields detected in the records list. This results in all columns being displayed in PowerBi, even if the value for the field in the first record is null.